% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{check_for_unique_id}
\alias{check_for_unique_id}
\title{Check for presence of unique ID}
\usage{
check_for_unique_id(data, vars, quiet = FALSE)
}
\arguments{
\item{data}{Data frame to be tested.}

\item{vars}{Variable names to check}

\item{quiet}{Whether to display messages in the console. Default is true.}
}
\value{
Character vector of the potential unique ID variable(s) names or "No
  ID" if none are found.
}
\description{
Checks `vars` to see if any are unique for each row. Returns the variable
name or names if more than 1 potential unique ID are found.
}
