% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{check_for_weight}
\alias{check_for_weight}
\title{Checks variable names for a potential weight variable}
\usage{
check_for_weight(weight = TRUE, data, vars, quiet = FALSE)
}
\arguments{
\item{weight}{Either a logical to check for a weight var, or the name of the
weight variable on which to run a summary. Default is \code{TRUE}.}

\item{data}{Data frame to be tested.}

\item{vars}{Character vector of variable names to be checked.}

\item{quiet}{Whether to display messages in the console. Default is true.}
}
\value{
Character vector of potential weight variable names or "No weight
variable found" if no potential weight variable identified.
}
\description{
Searches for a variable name that includes "weight" in the variable names
provided. If none found it will return "No weight variable found", otherwise
it will return the variable names that match.
}
