% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labels.R
\name{extract_labels}
\alias{extract_labels}
\title{Extract raw data meta data}
\usage{
extract_labels(loaded_data, tibble_out = TRUE)
}
\arguments{
\item{loaded_data}{A tibble created by reading in an SPSS .sav file using the
haven package. The data must contain SPSS metadata (variable and value labels).}

\item{tibble_out}{Logical indicating whether to return the output as a list
of tibbles (default is TRUE). If FALSE, the output will be returned as
data.tables.}
}
\value{
A list containing four elements:
  \itemize{
    \item var_labels: A tibble/data.table with variable names and their labels
      containing columns 'var_name' and 'var_label'
    \item val_labels: A tibble/data.table with value labels for categorical variables
      containing columns 'var_name', 'val_label', and 'val_value'
    \item no_var_labels: A tibble/data.table with names of variables that have no
      variable labels
    \item no_val_labels: A tibble/data.table with names of variables that have no
      value labels
  }
}
\description{
This function extracts all variable and value labels (i.e., metadata) from an
SPSS file. It also identifies the variables for which no metadata is available.
}
\details{
The function performs several operations:
\itemize{
  \item Extracts variable labels for all variables in the dataset
  \item Identifies variables that lack variable labels
  \item Extracts value labels for categorical variables
  \item Validates that no character variables have value labels (throws an error if found)
  \item Identifies variables that lack value labels
}

The function will throw an error if it encounters any character variables with
value labels, as these are not supported. The error message will list the
problematic variables that need to be checked.
}
\examples{
data(test_sav)
test_sav <- test_sav |>
  dplyr::select(-lab_var3)

# extract variable and value labels
labels_list <- extract_labels(test_sav)

}
