% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valLabels.R
\name{get_valLabels}
\alias{get_valLabels}
\title{Extract value labels from a tibble with haven_labelled columns}
\usage{
get_valLabels(datafile)
}
\arguments{
\item{datafile}{A tibble created by reading in an SPSS .sav file using the haven package}
}
\description{
Extract value labels from a tibble with haven_labelled columns
}
