% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varLabels.R
\name{get_varLabels}
\alias{get_varLabels}
\title{extract variable labels from a tibble with haven_labelled columns}
\usage{
get_varLabels(datafile)
}
\arguments{
\item{datafile}{A tibble created by reading in an SPSS .sav file using the haven package}
}
\description{
extract variable labels from a tibble with haven_labelled columns
}
