% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sav.R
\name{load_sav}
\alias{load_sav}
\title{Load tidy SPSS Data}
\usage{
load_sav(data_path, keyv = c("respid"), col_omit = NULL, tibble_out = TRUE)
}
\arguments{
\item{data_path}{Path to the SPSS data file.}

\item{keyv}{A character vector of variable names to set as keys for the data.table.}

\item{col_omit}{A character vector of column names to omit from the loaded data.}

\item{tibble_out}{Logical indicating whether to return the output as a list
of tibbles (default is TRUE). If FALSE, the output will be returned as
data.tables.}
}
\value{
A list containing:
\describe{
\item{var_labels}{A data.table or tibble containing variable labels.}
\item{no_var_labels}{A data.table or tibble containing variables with no variable labels.}
\item{val_labels}{A data.table or tibble containing value labels.}
\item{no_val_labels}{A data.table or tibble containing variables with no value labels.}
\item{loaded_data}{The loaded SPSS data as a data.table or tibble.}
}
}
\description{
This function loads SPSS data and creates data.tables for variable labels and value labels.
}
\details{
The loaded SPSS data has all labels removed via \code{haven::zap_labels()}.
}
\examples{
# Define path to SPSS file to be loaded
data_path <- system.file("extdata", "test_sav.sav", package = "tveDataLoader")

# Load the SPSS file using load_sav omitting non-valid columns
loaded_data <- load_sav(data_path, col_omit = 'lab_var3')

}
