% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_dat_val_labels}
\alias{new_dat_val_labels}
\title{Value Labels Dataset}
\format{
A tibble with 4 rows and 3 columns:
\describe{
  \item{variable}{Variable name in the survey dataset}
  \item{value}{Numeric value of the response}
  \item{value label}{Descriptive label for the response value}
}
}
\source{
Generated sample data for testing purposes
}
\usage{
new_dat_val_labels
}
\description{
A dataset containing value labels for categorical survey responses.
}
\keyword{datasets}
