% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{summarise_weight_var}
\alias{summarise_weight_var}
\title{Run summary statistics on weight variable}
\usage{
summarise_weight_var(weight_var, data, quiet = FALSE)
}
\arguments{
\item{weight_var}{Character vector of weight variable names in \code{data} to be summarised.}

\item{data}{Dataframe that contains \code{weight_var}(s).}

\item{quiet}{Whether to display messages in the console. Default is true.}
}
\value{
A tibble of summary statistics with 1 row per \code{weight_var}.
}
\description{
Run summary statistics on weight variable
}
