% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_dat}
\alias{test_dat}
\title{Test Dataset}
\format{
A tibble with 100 rows and 4 columns:
\describe{
  \item{respid}{Respondent ID (1-100)}
  \item{q1}{A labelled variable with values 1-4, labelled as A through D}
  \item{q1.1}{A labelled variable with values 1-4, labelled as A through D}
  \item{weight}{Survey weight, all set to 1}
}
}
\source{
Generated sample data for testing purposes
}
\usage{
test_dat
}
\description{
A sample dataset containing survey responses with labelled variables.
}
\keyword{datasets}
