% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_sav}
\alias{test_sav}
\title{Test SPSS Dataset}
\format{
A tibble with 100 rows and 6 columns:
\describe{
\item{respid}{A labelled variable with Respondent ID (1-100)}
\item{lab_var1}{A labelled variable with values 1-4, labelled as A through D}
\item{lab_var2}{A labelled variable with values 1-2, labelled as a and b}
\item{num_var1}{A numeric variable with random normal distribution}
\item{char_var1}{A character variable with random letters}
\item{lab_var3}{A labelled character variable with random letters}
}
}
\source{
Generated sample data for testing purposes
}
\usage{
test_sav
}
\description{
A sample dataset containing various types of variables to test SPSS data loading functionality.
This dataset is also available as a raw SPSS (.sav) file in the package's extdata directory.
}
\examples{
# Load the dataset directly
data(test_sav)

# Or load the SPSS file using read_sav
spss_path <- system.file("extdata", "test_sav.sav", package = "tveDataLoader")
test_sav_loaded <- haven::read_sav(spss_path)
}
\keyword{datasets}
