% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tve_path.R
\name{tve_path}
\alias{tve_path}
\title{Generate an OS specific TVE filepath}
\usage{
tve_path(path)
}
\arguments{
\item{path}{File path to be converted}
}
\description{
Paths on macOS don't use named drives, meaning a filepath from a Windows
machine won't work on macOS and vice versa. This function should translate
between the two, though primarily from Windows to Mac.
}
