% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_labels.R
\name{update_val_labels}
\alias{update_val_labels}
\title{Update value labels of a `labelled` tibble}
\usage{
update_val_labels(dat, new_val_labels)
}
\arguments{
\item{dat}{Tibble containing variables to be updated}

\item{new_val_labels}{Tibble with three columns, variable, value and value
label. i.e. the format generated by [tveDataLoader::get_valLabels()].
Column names don't matter, but the first column will be used as variable,
the second as value and the third as value label, so they must be in that
order.}
}
\value{
Version of dat with the updated variable labels applied
}
\description{
Updates the value labels of a tibble using
[labelled::set_value_labels()]. The values of the variables to be update should all have
a row in `new_val_labels`. Variables that don't won't be changed and will
keep their existing label.
}
\examples{
dat <- tibble::tibble(
  a = labelled::labelled(c(1, 2), labels = c("value 1" = 1, "value 2" = 2))
)
tveDataLoader::get_valLabels(dat)
new_val_labs <- tibble::tribble(
  ~variable, ~value, ~`value label`,
  "a", 1, "new value 1",
  "a", 2, "new value 2"
)
dat <- update_val_labels(dat, new_val_labs)
tveDataLoader::get_valLabels(dat)

}
