% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_labels.R
\name{update_var_labels}
\alias{update_var_labels}
\title{Update variable labels of a `labelled` tibble}
\usage{
update_var_labels(dat, new_var_labels)
}
\arguments{
\item{dat}{Tibble containing variables to be updated}

\item{new_var_labels}{Tibble with two columns, variable and label. i.e. the
format generated by [tveDataLoader::get_varLabels()]. Column names don't
matter, but the first column will be used as variable and the second as
label, so they must be in that order.}
}
\value{
Version of dat with the updated variable labels applied
}
\description{
Updates the variable labels of a tibble using
[labelled::set_variable_labels()]. The variables to be update should all have
a row in `new_var_labels`. Variables that don't won't be changed and will
keep their existing label.
}
\examples{
dat <- tibble(a = 1, b = 2)
tveDataLoader::get_varLabels(dat)
new_var_labs <- tibble::tribble(
  ~variable, ~label,
  "a", "variable 1",
  "b", "variable 2"
)
dat <- update_var_labels(dat, new_var_labs)
tveDataLoader::get_varLabels(dat)

}
