% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{check_for_vars}
\alias{check_for_vars}
\title{Checks for presence of `vars` in `data`}
\usage{
check_for_vars(data, vars, quiet = FALSE)
}
\arguments{
\item{data}{Data frame to be tested.}

\item{vars}{Character vector of variable names. `data` will be checked for
the presence of each variable}

\item{quiet}{Whether to display messages in the console. Default is true.}
}
\description{
Checks for presence of `vars` in `data`
}
