\name{get_varLabels}
\alias{get_varLabels}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extract variable labels from an SPSS .sav file read in with tveDataLoader::read_sav()
}
\usage{
get_varLabels(datafile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datafile}{
Object of class .tibble, with variable labels.
}
}
\value{
Object of class .tibble with 2 columns: variable and label
}
\references{
%% ~put references to the literature/web site here ~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
