% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sav.R
\name{read_sav}
\alias{read_sav}
\title{Read in SPSS .sav files and convert variable names to lower case}
\usage{
read_sav(filepath)
}
\arguments{
\item{filepath}{Character storing filepath, including filename, of the sav file to be read in}
}
\description{
Wrapper around haven::read_sav() and stringr::str_to_lower()
}
