% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{run_data_checks}
\alias{run_data_checks}
\title{Data checking functions}
\usage{
run_data_checks(
  data,
  vars = NULL,
  unique_id = TRUE,
  weight = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{data}{Data frame to be tested.}

\item{vars}{Optional character vector of variable names to be tested within
data}

\item{unique_id}{Logical. Whether to run the unique_id check. Default is
`TRUE`.}

\item{weight}{Either a logical to check for a weight var, or the name of the
weight variable on which to run a summary. Default is `TRUE`.}

\item{quiet}{Whether to display messages in the console. Default is true.}
}
\description{
Performs standard data checks on survey data. Doesn't alter the data in any
way, just reports on potential data quality concerns.
}
