% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{run_data_checks}
\alias{run_data_checks}
\title{Data checking functions}
\usage{
run_data_checks(
  data,
  vars = NULL,
  unique_id = TRUE,
  weight = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{data}{Data frame to be tested.}

\item{vars}{Optional character vector of variable names to be tested within
data. Default is NULL, i.e. test all variables.}

\item{unique_id}{Logical. Whether to run the unique_id check. Default is
`TRUE`.}

\item{weight}{Either a logical to check for a weight var, or the name of the
weight variable on which to run a summary. Default is `TRUE`.}

\item{quiet}{Whether to display messages in the console. Default is true.}
}
\value{
List of check outcomes:
 \itemize{
  \item{"data": `data` passed through}
  \item{"unique_id": Names of variables identified as possible unique identifiers}
  \item{"weight_var": Names of variables identified as possible weight variables}
  \item{"weight_summary": Tibble of summary stats on possible weight variables}
  \item{"vars_check": Character vector of variables in `vars` that are NOT present in `data`}
  }
}
\description{
Performs standard data checks on survey data. Doesn't alter the data in any
way, just reports on potential data quality concerns.
}
