% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tveDataLoader-package.R
\docType{package}
\name{tveDataLoader-package}
\alias{tveDataLoader}
\alias{tveDataLoader-package}
\title{tveDataLoader: DTAG Analytics Data Loading Services}
\description{
This package provides the functionality required by the DTAG Analytics team to load SPSS data into R. It includes data cleaning and will be expanded in the future to include other data formats.
}
\author{
\strong{Maintainer}: Juan Hernandez \email{juan.hernandez@thevalueengineers.com}

}
\keyword{internal}
